storage implemented

+++ internal/grpc/auth/server.go
@@ -21,7 +21,7 @@ const (
 
 type Auth interface {
 	Login(ctx context.Context, email string, password string, appID int) (token string, err error)
-	Register(ctx context.Context, email string, passwrd string) (userID int, err error)
+	Register(ctx context.Context, email string, passwrd string) (userID int64, err error)
 	Logout(token string) (success bool, err error)
 	IsAdmin(userId int) (isAdmin bool, err error)
 }
